+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/listener/BungeeCordListener.java
@@ -67,7 +67,6 @@ public class BungeeCordListener implements PluginMessageListener {
 
                     //put it only if the user doesn't has a session open
                     //so that the player have to send the bungeecord packet and cannot skip the verification then
-
                     plugin.getSessions().putIfAbsent(checkedPlayer.getAddress().toString(), playerSession);
                 } else {
                     //blacklist target for the current login